% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mash_table.R
\name{as.data.frame.Mash_table}
\alias{as.data.frame.Mash_table}
\title{Title}
\usage{
\method{as.data.frame}{Mash_table}(dat, mash_method = "row", ...)
}
\arguments{
\item{...}{parameters passed on to \code{as.data.frame.data.table}}
}
\description{
Stacking uses \code{data.table}s internally,
}

