% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_table.R
\name{as.data.table.Composite_table}
\alias{as.data.table.Composite_table}
\alias{as.data.frame.Composite_table}
\title{Convert a Composite Table to a data.table or data.frame}
\usage{
\method{as.data.table}{Composite_table}(x, multinames = TRUE, sep = ".",
  ...)

\method{as.data.frame}{Composite_table}(x, row.names = NULL,
  optional = FALSE, multinames = TRUE, sep = ".", ...)
}
\arguments{
\item{x}{a \code{Composite_table}}

\item{multinames}{logical. Whether to prepend multinames before the column
names}

\item{sep}{separator between multinames and individual column names}

\item{...}{ignored}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link{data.frame}(*, check.names = !optional)}.}
}
\value{
a \code{data.table} or \code{data.frame}
}
\description{
As a \code{Composite_table} already is a \code{data.table} this function
does very little except stripping all additional attributes and classes,
as well as offering you the option to prepend the \code{multinames} before
the column names
}
