% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mash_table.R
\name{as.data.table.Mash_table}
\alias{as.data.table.Mash_table}
\title{Title}
\usage{
\method{as.data.table}{Mash_table}(dat, mash_method = "row", ...)
}
\arguments{
\item{...}{passed on to as.data.frame.data.table}
}
\description{
Title
}
