% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack_table.R
\name{as.data.table.Stack_table}
\alias{as.data.table.Stack_table}
\title{Title}
\usage{
\method{as.data.table}{Stack_table}(dat, stack_method = "row", ...)
}
\arguments{
\item{...}{passed on to as.data.frame.data.table}
}
\description{
Title
}

