% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{as_latex.Tatoo_report}
\alias{as_latex.Tatoo_report}
\title{Convert a Tatoo Report to Latex Code}
\usage{
\method{as_latex}{Tatoo_report}(x, ...)
}
\arguments{
\item{x}{a \code{\link{Tatoo_table}}, \code{data.frame} or a list of \code{data.frame}s}

\item{...}{passed on to methods}
}
\value{
\code{as_latex()}returns a \code{character} scalar of Latex code

\code{save_pdf()} returns a the path to the saved file as \code{character}
scalar.

\code{view_pdf()} returns \code{NULL} (invisibly)
}
\description{
Convert a Tatoo Report to Latex Code
}
\seealso{
Other as_latex methods: \code{\link{as_latex.Composite_table}},
  \code{\link{as_latex.Mashed_table}},
  \code{\link{as_latex.Tagged_table}},
  \code{\link{as_latex.data.frame}}, \code{\link{as_latex}}
}
