% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mashed_table.R
\name{as_Mashed_table}
\alias{as_Mashed_table}
\alias{is_Mashed_table}
\title{Coerce to Mashed Table}
\usage{
as_Mashed_table(dat, ...)

is_Mashed_table(dat, ...)
}
\arguments{
\item{dat}{any R object}

\item{...}{\code{mash_table()} only: \code{data.frame}s with the same row and column
count. Elements of \code{(...)} can be named, but the name must differ from
the argument names of this function.}
}
\value{
\code{as_Mashed_table()} returns a Mashed_table

\code{is_Mashed_table()} returns \code{TRUE} if its argument is a Mashed_table
and \code{FALSE} otherwise.
}
\description{
Coerce to Mashed Table
}
