% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_table.R
\name{as_multinames}
\alias{as_multinames}
\title{Create Composite Table multinames from a character vector}
\usage{
as_multinames(x)
}
\arguments{
\item{x}{a character vector of equal length as the data.frame for which it
the multinames should be created.}
}
\value{
a named integer vector that can be used as multinames attribute
for a \link{Composite_table}
}
\description{
Create Composite Table multinames from a character vector
}
\examples{

dat <- data.frame(
  apple = 1,
  banana = 2,
  dog = 1,
  cat = 2,
  parrot = 3
)

multinames(dat) <- as_multinames(
  c('fruit', 'fruit', 'animal', 'animal', 'animal')
)

multinames(dat)

}
