% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workbook.R
\name{as_workbook.Mash_table}
\alias{as_workbook.Mash_table}
\title{As Workbook}
\usage{
\method{as_workbook}{Mash_table}(dat, mash_method, insert_blank_row,
  sep_height = 20, sheet_name = "sheet1", ...)
}
\arguments{
\item{...}{parameters passed on to openxlsx::writeDate}
}
\description{
As Workbook
}
