% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workbook.R
\name{as_workbook.Pub_report}
\alias{as_workbook.Pub_report}
\title{Convert Pub_report to openxlsx Workbook}
\usage{
\method{as_workbook}{Pub_report}(dat)
}
\arguments{
\item{dat}{a pub report}
}
\description{
Converts a Pub_report to a an \code{openxlsx::Workbook} object. Sheet names
will be taken from the element names of Pub_report
(\code{names(yourPub_report)}).
}

