% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workbook.R
\name{as_workbook.TT_report}
\alias{as_workbook.TT_report}
\title{Convert TT_report to openxlsx Workbook}
\usage{
\method{as_workbook}{TT_report}(dat)
}
\arguments{
\item{dat}{a pub report}
}
\description{
Converts a TT_report to a an \code{openxlsx::Workbook} object. Sheet names
will be taken from the element names of TT_report
(\code{names(yourTT_report)}).
}

