% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp_table.R
\name{comp_table_list}
\alias{comp_table_list}
\title{Composite Table}
\usage{
comp_table_list(tables, titles = names(tables), by = NULL, meta = NULL)
}
\arguments{
\item{titles}{}
}
\description{
Composite Table
}
