% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{default_kable_options}
\alias{default_kable_options}
\title{Default Kable options for as_latex and co}
\usage{
default_kable_options(...)
}
\arguments{
\item{...}{additional arguments added to the options list}
}
\description{
\code{default_kable_options()} returns a list of the default options that are
required for \code{\link[=as_latex]{as_latex()}} to work correctly. Those defaults should not be
modified, but you can pass additional \code{\link[knitr:kable]{knitr::kable()}} options to
\code{as_latex()} to modify the output a bit.
}
\examples{

default_kable_options

as_latex(iris, kable_options = default_kable_options(digits = 0))

}
