% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hammr.R
\name{df_typecast_all}
\alias{df_typecast_all}
\title{Typecast all columns of a data.frame of a specific type}
\usage{
df_typecast_all(dat, from = "factor", to = "character")
}
\arguments{
\item{dat}{a data.frame}

\item{from}{column type to cast}

\item{to}{target column type}
}
\value{
a data frame with all columns of class from converted to class to
}
\description{
Bulk convert columns of a data.frame that share a certain class to a different
class. Use with care, will introduce NAs for some conversion attempts
}
