% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hammr.R
\name{is_any_class}
\alias{is_any_class}
\title{Check if any of the classes of the object match a certain string}
\usage{
is_any_class(dat, choices)
}
\arguments{
\item{dat}{the object}

\item{choices}{the class to be checked for}
}
\value{
True if any of the object classes are the desired class
}
\description{
Check if any of the classes of the object match a certain string
}
