% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stacked_table.R
\name{is_Stacked_table}
\alias{is_Stacked_table}
\title{Test If Object is a Stacked_table}
\usage{
is_Stacked_table(x)
}
\arguments{
\item{x}{Any \R object.}
}
\value{
\code{is_Stacked_table()} returns \code{TRUE} if its argument is a \code{Stacked_table}
and \code{FALSE} otherwise.
}
\description{
Test If Object is a Stacked_table
}
