% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{is_Tagged_table}
\alias{is_Tagged_table}
\title{Check if object is a Tagged_table}
\usage{
is_Tagged_table(dat)
}
\arguments{
\item{dat}{any R object}
}
\description{
Check if object is a Tagged_table
}
