% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tatoo_report.R
\name{is_Tatoo_report}
\alias{is_Tatoo_report}
\title{Test if Object is a Tatoo_report}
\usage{
is_Tatoo_report(x)
}
\arguments{
\item{x}{Any \R object.}
}
\value{
\code{is_Tatoo_report()} returns \code{TRUE} if its argument is a \code{Tatoo_report}
and \code{FALSE} otherwise.
}
\description{
Test if Object is a Tatoo_report
}
