% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tatoo_table.R
\name{is_Tatoo_table}
\alias{is_Tatoo_table}
\title{Test if objects is a Tatoo_table}
\usage{
is_Tatoo_table(x)
}
\arguments{
\item{x}{Any \R object.}
}
\value{
\code{is_Tatoo_table} returns \code{TRUE} if its argument is a \code{Tatoo_table}
and \code{FALSE} otherwise.
}
\description{
Test if objects is a Tatoo_table
}
