% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hammr.R
\name{looks_like_integer}
\alias{looks_like_integer}
\title{Check if a value looks like integer}
\usage{
looks_like_integer(x, na_value = FALSE, scalar = FALSE)
}
\arguments{
\item{x}{...}

\item{na_value}{What should NAs be treated es (should be TRUE, FALSE or NA)}

\item{scalar}{allow only scalars (vectors of length 1)}
}
\value{
TRUE / FALSE
}
\description{
Check if a value looks like integer
}
