% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mash_table.R
\name{mash_table}
\alias{mash_table}
\alias{rmash}
\alias{cmash}
\title{Mash Table}
\usage{
mash_table(..., rem_ext = NULL)

rmash(..., rem_ext = NULL, insert_blank_row = FALSE)

cmash(..., rem_ext = NULL, by = NULL, suffixes = NULL)
}
\arguments{
\item{rem_ext}{}
}
\description{
Stack tables are designed to make it easy to put together multidimensional
tables from two data.frames. An example where this might be useful is
if you have a data.frame of numeric values, and a second data.frame of
associated standard errors.
}
\details{
Stack table provides a framework to stack those two data.frames together
into one data.frame with alternating rows or columns. You can then output the
stacked table as latex \code{\link{print_tex.Mash_table}}, as xlsx
\code{\link{save_as.Mash_table}} or simply as data.table or data.frame
via \code{as.data.table} or \code{as.data.frame}.

If your goal is to present formated tables as latex or xlsx, you  should aslo
look into \code{\link{df_format}}, \code{\link{df_round}} and
\code{\link{df_signif}}.
}
