% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_table.R
\name{multinames<-}
\alias{multinames<-}
\alias{multinames}
\title{Set the multinames attribute of a Composite_table}
\usage{
multinames(dat) <- value

multinames(dat)
}
\arguments{
\item{dat}{a Composite_table or data.frame}

\item{value}{a named vector of ascending integers. The name is the
multi-column heading, the integer value is the last column that this
heading applies to}
}
\description{
Set the multinames attribute of a Composite_table
}
\examples{

df_mean <- data.frame(
  Species = c("setosa", "versicolor", "virginica"),
  length = c(5.01, 5.94, 6.59),
  width = c(3.43, 2.77, 2.97)
)

multinames(df_mean) = c('species' = 1, measures = 3)

# .species..     ...measures...
# 1    Species     length   width
# 2     setosa       5.01    3.43
# 3 versicolor       5.94    2.77
# 4  virginica       6.59    2.97


}
\seealso{
\link{Composite_table}, \code{\link[=as_multinames]{as_multinames()}}
}
