% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{multinames_to_colspans}
\alias{multinames_to_colspans}
\title{Convert multinames to colspans}
\usage{
multinames_to_colspans(x)
}
\arguments{
\item{x}{a \link{Composite_table} \link{multinames} attribute.}
}
\value{
A named character vector of colspans (for \code{\link[kableExtra:add_header_above]{kableExtra::add_header_above()}})
}
\description{
Convert multinames to colspans
}
