% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_file}
\alias{open_file}
\title{Open a file}
\usage{
open_file(x)
}
\arguments{
\item{x}{\code{character} scalar. Path to the file to open.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Open a file with the default associated program. Might behave differently
depending on the operating system.
}
