% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mashed_table.R
\name{print.Mashed_table}
\alias{print.Mashed_table}
\title{Printing Mashed Tables}
\usage{
\method{print}{Mashed_table}(
  x,
  mash_method = attr(x, "mash_method"),
  insert_blank_row = attr(x, "insert_blank_row"),
  id_vars = attr(x, "id_vars"),
  ...
)
}
\arguments{
\item{x}{a Mashed_table}

\item{mash_method}{either \code{"row"} or \code{"col"}. Should the tables be mashed
together with alternating rows or with alternating columns?}

\item{insert_blank_row}{Only if mashing rows: logical. Whether to insert
blank rows between mash-groups. \emph{Warning: this converts all columns to
character.} Use with care.}

\item{id_vars}{Only if mashing columns: one ore more colnames of the tables
to be mashed. If supplied, columns of both input tables are combined with
\code{\link[=merge]{merge()}}, otherwise \code{\link[=cbind]{cbind()}} is used.}

\item{...}{passed on to \code{\link[=print]{print()}}}
}
\value{
\code{x} (invisibly)
}
\description{
Printing Mashed Tables
}
