% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stacked_table.R
\name{print.Stacked_table}
\alias{print.Stacked_table}
\title{Printing Stacked Tables}
\usage{
\method{print}{Stacked_table}(x, ...)
}
\arguments{
\item{x}{A \code{\link{Stacked_table}}}

\item{...}{passed on to \code{\link[=print]{print()}}}
}
\value{
\code{x} (invisibly)
}
\description{
Printing Stacked Tables
}
