% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{print.Tagged_table}
\alias{print.Tagged_table}
\title{Printing Tagged Tables}
\usage{
\method{print}{Tagged_table}(x, ...)
}
\arguments{
\item{x}{a \code{Tagged_table}}

\item{...}{passed on to \code{\link{print}}}
}
\value{
\code{x} (invisibly)
}
\description{
Printing Tagged Tables
}
