% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tatoo_report.R
\name{print.Tatoo_report}
\alias{print.Tatoo_report}
\title{Printing Tatoo Reports}
\usage{
\method{print}{Tatoo_report}(x, ...)
}
\arguments{
\item{x}{A \code{Tatoo_report}}

\item{...}{passed on to \code{\link{print}}}
}
\value{
\code{x} (invisibly)
}
\description{
Printing Tatoo Reports
}
