% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{print.TT_meta}
\alias{print.TT_meta}
\title{Printing Tagged Table Metadata}
\usage{
\method{print}{TT_meta}(x, ...)
}
\arguments{
\item{x}{A \code{TT_meta} object}

\item{...}{ignored}
}
\value{
\code{x} (invisibly)
}
\description{
Printing Tagged Table Metadata
}
