% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub_report.R
\name{pub_report}
\alias{pub_report}
\title{A list of Pub_tables}
\usage{
pub_report(dat)
}
\arguments{
\item{dat}{}
}
\value{
An object of class 'Pub_report'
}
\description{
A Pub report is a list of Pub_tables. Right now this class is mostly a
placeholder, but in the future it will support various export methods.
}
