% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub_table.R
\name{pub_table_meta}
\alias{pub_table_meta}
\title{Pub Table Metadata}
\usage{
pub_table_meta(table_id, title, longtitle = title, subtitle = NULL,
  footer = NULL, ...)
}
\arguments{
\item{table_id}{}

\item{...}{}
}
\description{
Create a pub table metadata object
}
