% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pub_table.R
\name{pub_tableMeta}
\alias{pub_tableMeta}
\title{Pub Table Metadata}
\usage{
pub_tableMeta(tableId, title, longtitle = title, subtitle = NULL, ...)
}
\arguments{
\item{tableId}{}

\item{...}{}
}
\description{
Create a pub table metadata object
}

