% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{regions}
\alias{regions}
\title{Get Named Regions of an Excel Sheet as Data.Table}
\usage{
regions(x)
}
\arguments{
\item{x}{An openxlsx workbook or a \code{character} vector with attributes
\code{position} and \code{sheet} as returned by \code{\link[openxlsx:getNamedRegions]{openxlsx::getNamedRegions()}}}
}
\value{
A \code{data.table}
}
\description{
Get Named Regions of an Excel Sheet as Data.Table
}
