% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xlsx.R
\name{save_xlsx.Mash_table}
\alias{save_xlsx.Mash_table}
\title{Title}
\usage{
\method{save_xlsx}{Mash_table}(dat, outfile, mash_method = "row",
  insert_blank_row = FALSE, sep_height = 24, overwrite = FALSE)
}
\arguments{
\item{...}{parameters passed on to}
}
\description{
Title
}

