% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workbook.R
\name{save_xlsx.StackTable}
\alias{save_xlsx.StackTable}
\title{Title}
\usage{
\method{save_xlsx}{StackTable}(dat, outfile, mash_method = "row",
  insert_blank_row = FALSE, sep_height = 24, sheet_name = "sheet1",
  overwrite = FALSE, ...)
}
\arguments{
\item{...}{parameters passed on to}
}
\description{
Title
}
