% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stacked_table.R
\name{spacing<-}
\alias{spacing<-}
\title{Set the spacing of a Stacked_table}
\usage{
spacing(x) <- value
}
\arguments{
\item{x}{a \code{Stacked_table}}

\item{value}{a scalar integer}
}
\description{
Set the number of lineskips between the tables when exporting to xlsx.
}
\seealso{
\link{Stacked_table}
}
