% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack_table.R
\name{stack_table}
\alias{cstack}
\alias{rstack}
\alias{stack_table}
\title{Stack Table}
\usage{
stack_table(dat1, dat2, rem_ext = NULL)

rstack(dat1, dat2, rem_ext = NULL, ...)

cstack(dat1, dat2, rem_ext = NULL, ...)
}
\arguments{
\item{rem_ext}{}
}
\description{
Stack tables are designed to make it easy to put together multidimensional
tables from two data.frames. An example where this might be useful is
if you have a data.frame of numeric values, and a second data.frame of
associated standard errors.
}
\details{
Stack table provides a framework to stack those two data.frames together
into one data.frame with alternating rows or columns. You can then output the
stacked table as latex \code{\link{print_tex.Stack_table}}, as xlsx
\code{\link{save_as.Stack_table}} or simply as data.table or data.frame
via \code{as.data.table} or \code{as.data.frame}.

If your goal is to present formated tables as latex or xlsx, you  should aslo
look into \code{\link{df_format}}, \code{\link{df_round}} and
\code{\link{df_signif}}.
}

