% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{meta<-}
\alias{meta<-}
\alias{meta}
\alias{table_id<-}
\alias{title<-}
\alias{longtitle<-}
\alias{subtitle<-}
\alias{footer<-}
\title{Set Tagged Table metadata}
\usage{
meta(dat) <- value

meta(dat)

table_id(dat) <- value

title(dat) <- value

longtitle(dat) <- value

subtitle(dat) <- value

footer(dat) <- value
}
\arguments{
\item{dat}{a \link{Tagged_table} or any object that can be converted to one}

\item{value}{value to assign.}
}
\description{
Convenience functions to modify Tagged_table metadata. If \code{dat} is not a
Tagged_table already, it will be converted to one.
}
\seealso{
\link{Tagged_table}, \link{tt_meta}
}
