% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tatoo_table.R
\name{tatoo_table}
\alias{tatoo_table}
\alias{Tatoo_table}
\title{Tatoo Table}
\usage{
tatoo_table(dat)
}
\arguments{
\item{dat}{an object of any of the classes listed in the description}
}
\description{
\code{Tatto_table} is the superclass of all the \code{*_table} classes made available
by this package. Each \code{Tatoo_table} provides a different way of combining
several tables (data.frames) into a single table. Those tables can then
be exported via \code{\link[=as_workbook]{as_workbook()}}/\code{\link[=save_xlsx]{save_xlsx()}}. In the future, support for
latex and html export is also planned.
}
\details{
Currently, the following subclasses exists:
\itemize{
\item \code{\link{Tagged_table}}
\item \code{\link{Composite_table}}
\item \code{\link{Mashed_table}}
\item \code{\link{Stacked_table}}
}

The \code{tatoo_table()} function is just a constructor used internally and you
will not need to use it except if your planning on extending this package
with your own code.
}
\seealso{
Other Tatoo tables: \code{\link{comp_table}},
  \code{\link{mash_table}}, \code{\link{stack_table}},
  \code{\link{tag_table}}
}
\concept{Tatoo tables}
