% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT_report.R
\name{tt_report}
\alias{tt_report}
\title{A list of Meta_tables}
\usage{
tt_report(dat)
}
\arguments{
\item{dat}{}
}
\value{
An object of class 'TT_report'
}
\description{
A Pub report is a list of Meta_tables. Right now this class is mostly a
placeholder, but in the future it will support various export methods.
}
