% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hammr.R
\name{vec_prioritise}
\alias{vec_prioritise}
\title{Rearrange vector based on priorities}
\usage{
vec_prioritise(x, high = NULL, low = NULL)
}
\arguments{
\item{x}{a character vector}

\item{high}{elements to be put to the front}

\item{low}{elements to be put to the back}
}
\value{
a reordered vector
}
\description{
Shoves elements of a character vector to the front or back.
Throws a warning if any elements of `high` or `low` are not present in `x`.
}
