% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_worksheet.R
\name{write_worksheet.Comp_table}
\alias{write_worksheet.Comp_table}
\title{Title}
\usage{
\method{write_worksheet}{Comp_table}(dat, wb, sheet, append = FALSE,
  start_row = 1L, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
