% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_worksheet.R
\name{write_worksheet.Pub_table}
\alias{write_worksheet.Pub_table}
\title{Write worksheet}
\usage{
\method{write_worksheet}{Pub_table}(dat, wb,
  sheet = sanitize_excel_sheet_names(attr(dat, "meta")$tableId))
}
\arguments{
\item{sheet}{}
}
\description{
Write worksheet
}

