% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_table.R
\name{as_Composite_table}
\alias{as_Composite_table}
\alias{is_Composite_table}
\title{Coerce to Composite Table}
\usage{
as_Composite_table(dat, id_vars, meta)

is_Composite_table(dat, ...)
}
\arguments{
\item{dat}{any R object}

\item{id_vars}{If id_vars is specified, the tables will be combined
using \code{\link[=merge]{merge()}} on the columns specified in id_vars, otherwise
the tables will be combined with \code{\link[=cbind]{cbind()}}.}

\item{meta}{a \link{TT_meta} object. If speciefied, the resulting
Composite_table will be wrapped in a \link{Tagged_table}.}

\item{...}{\code{comp_table} only: individual data.frames. A name can be
provided for each data.frame that will be used by \link{print} and
\code{\link[=as_workbook]{as_workbook()}} to create multi-table headings.}
}
\value{
\code{as_Composte_table()} returns a Composite_table

\code{is_Composite_table()} returns \code{TRUE} if its argument is a Composite_table
and \code{FALSE} otherwise.
}
\description{
Coerce to Composite Table
}
