% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{assign_tt_meta}
\alias{assign_tt_meta}
\title{Assign tt_meta elements}
\usage{
assign_tt_meta(x, assignment)
}
\arguments{
\item{x}{a \link{Tatoo_table} or data.frame}

\item{assignment}{A named list of length one, for example
\code{list(longtitle = value)}}
}
\description{
Internal function used by the metadata set functions
}
