% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_table.R
\name{flip_names}
\alias{flip_names}
\title{Flip names and multinames of a Composite Table}
\usage{
flip_names(dat, id_vars)
}
\arguments{
\item{dat}{A \code{Composite_table}}

\item{id_vars}{a character vector of column names of \code{dat}. The selected
columns will not be sorted lexically but kept to the left. If the columns
have a multiname associated with them, they must be supplied in the format
\code{column_name.multiname}.}
}
\value{
a Composite_table
}
\description{
The column names of the resulting \code{Composite_table} will be sorted
lexically
}
\examples{

dat <- comp_table(
  cars1 = head(cars),
  cars2 = tail(cars),
  data.frame(id = LETTERS[1:6])
)

flip_names(dat)
flip_names(dat, id_vars = "id")
flip_names(dat, id_vars = c("id", "speed.cars1"))

}
