% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stacked_table.R
\name{is_Stacked_table}
\alias{is_Stacked_table}
\title{Check if object is a Stacked_table}
\usage{
is_Stacked_table(dat)
}
\arguments{
\item{dat}{any R object}
}
\description{
Check if object is a Stacked_table
}
