% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{is_Tagged_table}
\alias{is_Tagged_table}
\title{Test If Object is a Tagged_table}
\usage{
is_Tagged_table(x)
}
\arguments{
\item{x}{Any \R object.}
}
\description{
Test If Object is a Tagged_table
}
