% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mashed_table.R
\name{mash_method<-}
\alias{mash_method<-}
\alias{insert_blank_row<-}
\alias{sep_height<-}
\alias{id_vars<-}
\title{Set mash attributes of a Mashed Table}
\usage{
mash_method(x) <- value

insert_blank_row(x) <- value

sep_height(x) <- value

id_vars(x) <- value
}
\arguments{
\item{x}{a \code{Mashed_table}}

\item{value}{a value that is legal for the individual attribute, as
described in \link{Mashed_table}}
}
\description{
Set mash attributes of a Mashed Table
}
\seealso{
\link{Mashed_table}
}
