% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Composite_table.R
\name{print.Composite_table}
\alias{print.Composite_table}
\title{Printing Composite Tables}
\usage{
\method{print}{Composite_table}(x, right = FALSE, ...)
}
\arguments{
\item{x}{a \code{Composite_table}}

\item{right}{Logical. Should strings be right aligned? The default is
left-alignment (the opposite of the standard \code{\link[=print.data.frame]{print.data.frame()}}).}

\item{...}{passed on to \code{\link{print}}}
}
\value{
\code{x} (invisibly)
}
\description{
Printing Composite Tables
}
