% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack_table.R
\name{stack_table}
\alias{stack_table}
\title{Stack tables}
\usage{
stack_table(..., meta = NULL)
}
\arguments{
\item{...}{A list whos elements can of class
\code{\link{Meta_table}},
\code{\link{Mash_table}},
\code{\link{Comp_table}},
or anything that can be coerced to a \code{data.frame} with
\code{as.data.frame}}

\item{meta}{a \code{\link{tt_meta}} object (optional)}
}
\description{
side by side or on top of each others
}

