% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{meta<-}
\alias{meta<-}
\alias{meta}
\alias{table_id<-}
\alias{title<-}
\alias{longtitle<-}
\alias{subtitle<-}
\alias{footer<-}
\title{Set Tagged Table metadata}
\usage{
meta(x) <- value

meta(x)

table_id(x) <- value

title(x) <- value

longtitle(x) <- value

subtitle(x) <- value

footer(x) <- value
}
\arguments{
\item{x}{a \code{\link{Tagged_table}} or any \R object that can be converted to one}

\item{value}{value to assign.}
}
\description{
Convenience functions to modify \code{Tagged_table} metadata. If \code{x} is not a
\code{Tagged_table} already, it will be converted to one.
}
\seealso{
\link{Tagged_table}, \link{tt_meta}
}
