% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tatoo-package.R
\docType{package}
\name{tatoo}
\alias{tatoo}
\alias{tatoo-package}
\alias{tatoo-package}
\title{tatoo: Combine and Export Data Frames}
\description{
tatoo: Combine and Export Data Frames
}
\section{Functions}{

\itemize{
\item \code{\link[=tag_table]{tag_table()}}: add captioning (title, footer, ...) to a table
\item \code{\link[=comp_table]{comp_table()}}: like \code{\link[=cbind]{cbind()}} or \code{\link[=merge]{merge()}}, but retain multi-column
headings
\item \code{\link[=mash_table]{mash_table()}}: combine data.frames so that their
rows or columns alternate. Mash tables are stored as lists that can be
converted to data.tables, or you can use \code{\link[=rmash]{rmash()}} and \code{\link[=cmash]{cmash()}} to create
\code{data.frames} directly.
\item \code{\link[=stack_table]{stack_table()}}: create a list of tables that can be exported to xlsx, all
tables on the same worksheet on top of each others
\item \code{\link[=compile_report]{compile_report()}}: create a list of tables that can be exported
to xlsx, one table per worksheet (a Stacked_table also counts as one table)
\item \code{\link[=as_workbook]{as_workbook()}} / \code{\link[=save_xlsx]{save_xlsx()}}: To export any of the objects described
above to excel workbooks.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statistikat/tatoo}
  \item Report bugs at \url{https://github.com/statistikat/tatoo/issues}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com}

}
