% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tatoo_report.R
\name{compile_report}
\alias{compile_report}
\alias{Tatoo_report}
\alias{tatoo_report}
\alias{compile_report_list}
\alias{is_Tatoo_report}
\title{Compile tables into a report}
\usage{
compile_report(...)

compile_report_list(dat)

is_Tatoo_report(dat, ...)
}
\arguments{
\item{...}{for compile_table\code{: individual Tatoo_table} or
data.frame` objects}

\item{dat}{for \code{compile_table_list}: A list of containing either
\code{Tatoo_table} or \code{data.frame} objects.}
}
\value{
A \code{Tatoo_report}: A list whose elements are either \code{data.frame}s
or \code{\link{Tatoo_table}}s
}
\description{
Compiles tables into a \code{Tatoo_report}. A \code{Tatoo_report} is just
a simple list object, but with special \code{print}, \code{as_workbook},
and \code{save_xlsx} methods. This makes it easy to save an arbitrary
number of tables to a single Excel workbook.
}
