% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tagged_table.R
\name{tt_meta}
\alias{tt_meta}
\alias{TT_meta}
\title{Tagged Table metadata}
\usage{
tt_meta(table_id = NULL, title = NULL, longtitle = title,
  subtitle = NULL, footer = NULL)
}
\arguments{
\item{table_id}{A scalar (will be coerced to \code{character})}

\item{title}{A scalar (will be coerced to \code{character})}

\item{longtitle}{A vector. If \code{length > 1} the title will be displayed
in several rows}

\item{subtitle}{A vector. If \code{length > 1} the title will be displayed
in several rows}

\item{footer}{A vector. If \code{length > 1} the title will be displayed
in several rows}
}
\value{
a TT_meta object.
}
\description{
Create a \code{TT_meta} (tagged table metadata) object. In the future,
different styling will be supported for title, longtitle and subtitle to
make the distinction more meaningful.
}
\seealso{
\link{Tagged_table}
}
