% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_table.R
\name{tt_meta}
\alias{tt_meta}
\title{Pub Table Metadata}
\usage{
tt_meta(table_id, title, longtitle = title, subtitle = NULL,
  footer = NULL, ...)
}
\arguments{
\item{table_id}{}

\item{...}{}
}
\description{
Create a pub table metadata object
}

