% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_worksheet.R
\name{write_worksheet}
\alias{write_worksheet}
\title{Title}
\usage{
write_worksheet(dat, wb, sheet, append = FALSE, start_row = 1L, ...)
}
\arguments{
\item{passed}{on to \code{openxlsx::writeData}}
}
\description{
Title
}
