% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_worksheet.R
\name{write_worksheet}
\alias{write_worksheet}
\title{Write data to an openxlsx Worksheet}
\usage{
write_worksheet(dat, wb, sheet, append = FALSE, start_row = 1L, ...)
}
\arguments{
\item{dat}{A tatool table object, \code{Comp_table}, \code{Meta_table},
\code{Mash_table} or \code{Stack_table}.}

\item{wb}{A \code{\link[openxlsx]{openxlsx}} Workbook object}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{append}{Logical. Whether or not to append to an exisiting worksheet or
create a new one}

\item{start_row}{A scalar specifiying the starting row to write to.}

\item{...}{passed onto methods}
}
\description{
This function is similar to \code{\link[openxlsx]{writeData}}, but
rather than writing \code{data.frames}, \code{write_data} supports methods
for the various \code{tatool} classes.
}
